Config recovery: write the comments for a reader who wasn't here
The comments were written in the vocabulary of the work rather than of
the code. "B1"/"B2", section numbers and vector numbers are all useful
to someone who has the planning documents open and meaningless to
everyone else, and a comment citing a document nobody can open leaves
the local obligation unrecorded - the reader sees an explanation where
there isn't one.

Where a citation was carrying the reasoning, the reasoning is now
stated. "Reuses §5.5's one-hour interval" becomes why an hour: the cost
being bounded is an extra namespace read per poll for a group likely to
keep failing, and the repair is not time-critical because the keys have
already been missing long enough to expire.

Folded the force rekey into ConfigRecovery as an extension, renamed to
match the repo's X+Y.swift convention for extension files. It keeps the
property that mattered - no dependency from the backfill to the rekey,
so the rekey deletes without touching the backfill - which was always
carried by the structure rather than by being a separate type.

Dropped the prose describing that removability. It documented a decision
that has not been made, so if the rekey is kept it becomes false with
nothing to catch it, and the structure already says it.

Also dropped a guard comment restating the doc comment three lines
above it, and corrected a MARK naming a function that had been renamed.

No behaviour change: 2277 passed / 0 failed.
